+++ src/components/GhostList.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
+import { ghostData } from '../data/ghostData'
+import styles from '../styles/GhostList.module.scss'
 import Ghost from './Ghost'
-import { ghostData } from './data/ghostData'
-import styles from './styles/GhostList.module.scss'
 
 const GhostList = () => {
   const ghostNodes = ghostData.map(ghost => {
