+++ src/components/GithubMark.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { useRecoilValue } from 'recoil'
-import githubLight from './static/GitHub-Mark-Light-64px.png'
-import githubDark from './static/GitHub-Mark-64px.png'
-import { darkModeState } from './utils/state'
-import styles from './styles/GithubMark.module.scss'
+import styles from '../styles/GithubMark.module.scss'
+import githubLight from '../static/GitHub-Mark-Light-64px.png'
+import githubDark from '../static/GitHub-Mark-64px.png'
+import { darkModeState } from '../utils/state'
 
 const GithubMark = () => {
   const darkMode = useRecoilValue(darkModeState)
