+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { RecoilRoot } from 'recoil'
-import App from './App'
+import App from './components/App'
 
 ReactDOM.render(
   <React.StrictMode>
